Rework Oracle tests to use BaseConnectorTest
