+++ plugin/trino-oracle/src/test/java/io/trino/plugin/oracle/TestOraclePoolConnectorTest.java
@@ -13,24 +13,23 @@
  */
 package io.trino.plugin.oracle;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.airlift.testing.Closeables;
 import io.trino.testing.QueryRunner;
 import io.trino.testing.sql.SqlExecutor;
+import io.trino.tpch.TpchTable;
 import org.testng.annotations.AfterClass;
+import org.testng.annotations.Test;
 
 import java.io.IOException;
 
 import static io.trino.plugin.oracle.TestingOracleServer.TEST_PASS;
 import static io.trino.plugin.oracle.TestingOracleServer.TEST_USER;
-import static io.trino.tpch.TpchTable.CUSTOMER;
-import static io.trino.tpch.TpchTable.NATION;
-import static io.trino.tpch.TpchTable.ORDERS;
-import static io.trino.tpch.TpchTable.REGION;
+import static io.trino.testing.sql.TestTable.randomTableSuffix;
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class TestRemarksReportingOraclePoolIntegrationSmokeTest
-        extends BaseOracleIntegrationSmokeTest
+public class TestOraclePoolConnectorTest
+        extends BaseOracleConnectorTest
 {
     private TestingOracleServer oracleServer;
 
@@ -48,9 +47,9 @@ public class TestRemarksReportingOraclePoolIntegrationSmokeTest
                         .put("connection-password", TEST_PASS)
                         .put("allow-drop-table", "true")
                         .put("oracle.connection-pool.enabled", "true")
-                        .put("oracle.remarks-reporting.enabled", "true")
+                        .put("oracle.remarks-reporting.enabled", "false")
                         .build(),
-                ImmutableList.of(CUSTOMER, NATION, ORDERS, REGION));
+                TpchTable.getTables());
     }
 
     @AfterClass(alwaysRun = true)
@@ -59,6 +58,20 @@ public class TestRemarksReportingOraclePoolIntegrationSmokeTest
     {
         Closeables.closeAll(oracleServer);
         oracleServer = null;
+    }
+
+    @Test
+    @Override
+    public void testCommentColumn()
+    {
+        String tableName = "test_comment_column_" + randomTableSuffix();
+
+        assertUpdate("CREATE TABLE " + tableName + "(a integer)");
+
+        // comment set
+        assertUpdate("COMMENT ON COLUMN " + tableName + ".a IS 'new comment'");
+        // without remarksReporting Oracle does not return comments set
+        assertThat((String) computeActual("SHOW CREATE TABLE " + tableName).getOnlyValue()).doesNotContain("COMMENT 'new comment'");
     }
 
     @Override
