+++ plugin/trino-oracle/src/test/java/io/trino/plugin/oracle/TestRemarksReportingOraclePoolConnectorTest.java
@@ -13,24 +13,20 @@
  */
 package io.trino.plugin.oracle;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.airlift.testing.Closeables;
 import io.trino.testing.QueryRunner;
 import io.trino.testing.sql.SqlExecutor;
+import io.trino.tpch.TpchTable;
 import org.testng.annotations.AfterClass;
 
 import java.io.IOException;
 
 import static io.trino.plugin.oracle.TestingOracleServer.TEST_PASS;
 import static io.trino.plugin.oracle.TestingOracleServer.TEST_USER;
-import static io.trino.tpch.TpchTable.CUSTOMER;
-import static io.trino.tpch.TpchTable.NATION;
-import static io.trino.tpch.TpchTable.ORDERS;
-import static io.trino.tpch.TpchTable.REGION;
 
-public class TestOraclePoolIntegrationSmokeTest
-        extends BaseOracleIntegrationSmokeTest
+public class TestRemarksReportingOraclePoolConnectorTest
+        extends BaseOracleConnectorTest
 {
     private TestingOracleServer oracleServer;
 
@@ -48,9 +44,9 @@ public class TestOraclePoolIntegrationSmokeTest
                         .put("connection-password", TEST_PASS)
                         .put("allow-drop-table", "true")
                         .put("oracle.connection-pool.enabled", "true")
-                        .put("oracle.remarks-reporting.enabled", "false")
+                        .put("oracle.remarks-reporting.enabled", "true")
                         .build(),
-                ImmutableList.of(CUSTOMER, NATION, ORDERS, REGION));
+                TpchTable.getTables());
     }
 
     @AfterClass(alwaysRun = true)
